feat(swap): adjust swap amount for gas fees, Introduce finalize swap transaction data and  useSendSwapTxMutation hook.
- Add useSendSwapTxMutation hook to encapsulate swap transaction sending logic
- Implement finalizeSwapTxData function to create immutable SwapTransactionData object
  - Subtract gas fees from swap amount if amount left in balance can not cover gas fees
  - Set amountAdjustedForFees flag to indicate if adjustment occurred
- Update SwapTool component to use useSendSwapTxMutation hook and finalizeSwapTxData function
- Define SwapTransactionData interface for structured swap transaction data
