+++ packages/web/hooks/swap/use-swap-assets.ts
@@ -1,0 +1,252 @@
+import type { Asset } from "@osmosis-labs/server";
+import { makeMinimalAsset } from "@osmosis-labs/utils";
+import { useRouter } from "next/router";
+import { useEffect, useMemo, useState } from "react";
+
+import { RecommendedSwapDenoms } from "~/config";
+import { AssetLists } from "~/config/generated/asset-lists";
+import { useDebouncedState } from "~/hooks/use-debounced-state";
+import { useShowPreviewAssets } from "~/hooks/use-show-preview-assets";
+import { useWalletSelect } from "~/hooks/wallet-select";
+import { useQueryParamState } from "~/hooks/window/use-query-param-state";
+import { useStore } from "~/stores";
+import { api } from "~/utils/trpc";
+
+/** Use assets for swapping: the from and to assets, as well as the list of
+ *  swappable assets. Sorts by balance if user is signed in.
+ *
+ *  Features:
+ *  * Option to store from and to assets in query params, default = yes
+ *  * Paginated swappable assets, with user balances if wallet connected
+ *  * Assets search query */
+export function useSwapAssets({
+  initialFromDenom = "ATOM",
+  initialToDenom = "OSMO",
+  useQueryParams = true,
+  useOtherCurrencies = true,
+} = {}) {
+  const { chainStore, accountStore } = useStore();
+  const account = accountStore.getWallet(chainStore.osmosis.chainId);
+  const { isLoading: isLoadingWallet } = useWalletSelect();
+
+  const {
+    fromAssetDenom,
+    toAssetDenom,
+    setFromAssetDenom,
+    setToAssetDenom,
+    switchAssets,
+  } = useToFromDenoms(useQueryParams, initialFromDenom, initialToDenom);
+
+  // generate debounced search from user inputs
+  const [assetsQueryInput, setAssetsQueryInput] = useState<string>("");
+  const [debouncedSearchInput, setDebouncedSearchInput] =
+    useDebouncedState<string>("", 500);
+  useEffect(() => {
+    setDebouncedSearchInput(assetsQueryInput);
+  }, [setDebouncedSearchInput, assetsQueryInput]);
+
+  const queryInput = useMemo(
+    () => (debouncedSearchInput ? { query: debouncedSearchInput } : undefined),
+    [debouncedSearchInput]
+  );
+
+  const { showPreviewAssets } = useShowPreviewAssets();
+
+  const canLoadAssets =
+    !isLoadingWallet &&
+    Boolean(fromAssetDenom) &&
+    Boolean(toAssetDenom) &&
+    useOtherCurrencies;
+  // use a separate query for search to maintain pagination in other infinite query
+  const {
+    data: selectableAssetPages,
+    isLoading: isLoadingSelectAssets,
+    fetchNextPage,
+    hasNextPage,
+    isFetchingNextPage,
+  } = api.edge.assets.getUserAssets.useInfiniteQuery(
+    {
+      search: queryInput,
+      userOsmoAddress: account?.address,
+      includePreview: showPreviewAssets,
+      limit: 50, // items per page
+    },
+    {
+      enabled: canLoadAssets,
+      getNextPageParam: (lastPage) => lastPage.nextCursor,
+      initialCursor: 0,
+    }
+  );
+
+  const allSelectableAssets = useMemo(
+    () => selectableAssetPages?.pages.flatMap(({ items }) => items) ?? [],
+    [selectableAssetPages?.pages]
+  );
+
+  const { asset: fromAsset } = useSwapAsset(
+    fromAssetDenom,
+    allSelectableAssets
+  );
+  const { asset: toAsset } = useSwapAsset(toAssetDenom, allSelectableAssets);
+
+  const recommendedAssets = useRecommendedAssets(
+    fromAsset?.coinMinimalDenom,
+    toAsset?.coinMinimalDenom
+  );
+
+  /** Remove to and from assets from assets that can be selected. */
+  const filteredSelectableAssets = useMemo(
+    () =>
+      allSelectableAssets.filter(
+        (asset) =>
+          asset.coinMinimalDenom !== fromAsset?.coinMinimalDenom &&
+          asset.coinMinimalDenom !== toAsset?.coinMinimalDenom
+      ) ?? [],
+    [allSelectableAssets, fromAsset, toAsset]
+  );
+
+  return {
+    fromAsset,
+    toAsset,
+    assetsQueryInput,
+    selectableAssets: filteredSelectableAssets,
+    isLoadingSelectAssets,
+    hasNextPageAssets: hasNextPage,
+    isFetchingNextPageAssets: isFetchingNextPage,
+    /** Recommended assets, with to and from tokens filtered. */
+    recommendedAssets,
+    setAssetsQueryInput,
+    setFromAssetDenom,
+    setToAssetDenom,
+    switchAssets,
+    fetchNextPageAssets: fetchNextPage,
+  };
+}
+
+/** Gets recommended assets directly from asset list. */
+function useRecommendedAssets(
+  fromCoinMinimalDenom?: string,
+  toCoinMinimalDenom?: string
+) {
+  return useMemo(
+    () =>
+      RecommendedSwapDenoms.map((denom) => {
+        const asset = AssetLists.flatMap(({ assets }) => assets).find(
+          (asset) => asset.symbol === denom
+        );
+        if (!asset) return;
+
+        return makeMinimalAsset(asset);
+      })
+        .filter((c): c is NonNullable<typeof c> => !!c)
+        .filter(
+          (currency) =>
+            currency &&
+            currency.coinMinimalDenom !== fromCoinMinimalDenom &&
+            currency.coinMinimalDenom !== toCoinMinimalDenom
+        ),
+    [fromCoinMinimalDenom, toCoinMinimalDenom]
+  );
+}
+
+/** Will query for an individual asset of any type of denom (symbol, min denom)
+ *  if it's not already in the list of existing assets. */
+function useSwapAsset<TAsset extends Asset>(
+  minDenomOrSymbol?: string,
+  existingAssets: TAsset[] = []
+) {
+  /** If `coinDenom` or `coinMinimalDenom` don't yield a result, we
+   *  can fall back to the getAssets query which will perform
+   *  a more comprehensive search. */
+  const existingAsset = existingAssets.find(
+    (asset) =>
+      asset.coinDenom === minDenomOrSymbol ||
+      asset.coinMinimalDenom === minDenomOrSymbol
+  );
+  !existingAsset;
+  const asset = useMemo(() => {
+    if (existingAsset) return existingAsset;
+
+    const asset = AssetLists.flatMap(({ assets }) => assets).find(
+      (asset) =>
+        minDenomOrSymbol &&
+        (asset.symbol === minDenomOrSymbol ||
+          asset.coinMinimalDenom === minDenomOrSymbol)
+    );
+    if (!asset) return;
+
+    return makeMinimalAsset(asset);
+  }, [minDenomOrSymbol, existingAsset]);
+
+  return {
+    asset: existingAsset ?? (asset as TAsset),
+  };
+}
+
+/** Switches on query params or react state to store to/from asset denoms.
+ *  The initial denoms will be ignored if the user set preferences via query params. */
+function useToFromDenoms(
+  useQueryParams: boolean,
+  initialFromDenom?: string,
+  initialToDenom?: string
+) {
+  const router = useRouter();
+
+  // user query params as state source-of-truth
+  // ignores initial denoms if there are query params
+  const [fromDenomQueryParam, setFromDenomQueryParam] = useQueryParamState(
+    "from",
+    useQueryParams ? initialFromDenom : undefined
+  );
+  const fromDenomQueryParamStr =
+    typeof fromDenomQueryParam === "string" ? fromDenomQueryParam : undefined;
+  const [toAssetQueryParam, setToAssetQueryParam] = useQueryParamState(
+    "to",
+    useQueryParams ? initialToDenom : undefined
+  );
+  const toDenomQueryParamStr =
+    typeof toAssetQueryParam === "string" ? toAssetQueryParam : undefined;
+
+  // if query params are not used, use react state as source-of-truth
+  const [fromAssetState, setFromAssetState] = useState<string | undefined>(
+    initialFromDenom
+  );
+  const [toAssetState, setToAssetState] = useState<string | undefined>(
+    initialToDenom
+  );
+
+  useEffect(() => {
+    setToAssetState(initialToDenom);
+    setFromAssetState(initialFromDenom);
+  }, [initialFromDenom, initialToDenom]);
+
+  // if using query params perform one push instead of two as the router
+  // doesn't handle two immediate pushes well within `useQueryParamState` hooks
+  const switchAssets = () => {
+    if (useQueryParams) {
+      const existingParams = router.query;
+      router.push({
+        query: {
+          ...existingParams,
+          from: toDenomQueryParamStr,
+          to: fromDenomQueryParamStr,
+        },
+      });
+      return;
+    }
+
+    const temp = fromAssetState;
+    setFromAssetState(toAssetState);
+    setToAssetState(temp);
+  };
+
+  return {
+    fromAssetDenom: useQueryParams ? fromDenomQueryParamStr : fromAssetState,
+    toAssetDenom: useQueryParams ? toDenomQueryParamStr : toAssetState,
+    setFromAssetDenom: useQueryParams
+      ? setFromDenomQueryParam
+      : setFromAssetState,
+    setToAssetDenom: useQueryParams ? setToAssetQueryParam : setToAssetState,
+    switchAssets,
+  };
+}
