+++ packages/web/hooks/swap/use-swap-query-router-best-quote.ts
@@ -1,0 +1,120 @@
+import { RouterKey } from "@osmosis-labs/server";
+import { createTRPCReact } from "@trpc/react-query";
+import { useMemo } from "react";
+
+import { useFeatureFlags } from "~/hooks/use-feature-flags";
+import { AppRouter } from "~/server/api/root-router";
+import { RouterInputs } from "~/utils/trpc";
+
+/** Iterates over available and identical routers and sends input to each one individually.
+ *  Results are reduced to best result by out amount.
+ *  Also returns the number of routers that have fetched and errored. */
+export function useQueryRouterBestQuote(
+  input: Omit<
+    RouterInputs["local"]["quoteRouter"]["routeTokenOutGivenIn"],
+    "preferredRouter"
+  >,
+  enabled: boolean,
+  routerKeys = ["legacy", "sidecar", "tfm"] as RouterKey[]
+) {
+  const featureFlags = useFeatureFlags();
+  const availableRouterKeys: RouterKey[] = useMemo(
+    () =>
+      !featureFlags._isInitialized
+        ? []
+        : routerKeys.filter((key) => {
+            if (!featureFlags.sidecarRouter && key === "sidecar") return false;
+            if (!featureFlags.legacyRouter && key === "legacy") return false;
+            // TFM doesn't support force swap through pool
+            if ((!featureFlags.tfmRouter || input.forcePoolId) && key === "tfm")
+              return false;
+            return true;
+          }),
+    [
+      featureFlags._isInitialized,
+      featureFlags.sidecarRouter,
+      featureFlags.legacyRouter,
+      featureFlags.tfmRouter,
+      routerKeys,
+      input.forcePoolId,
+    ]
+  );
+
+  const trpcReact = createTRPCReact<AppRouter>();
+  const routerResults = trpcReact.useQueries((t) =>
+    availableRouterKeys.map((key) =>
+      t.local.quoteRouter.routeTokenOutGivenIn(
+        {
+          ...input,
+          preferredRouter: key,
+        },
+        {
+          enabled: enabled && Boolean(availableRouterKeys.length),
+
+          // quotes should not be considered fresh for long, otherwise
+          // the gas simulation will fail due to slippage and the user would see errors
+          staleTime: 5_000,
+          cacheTime: 5_000,
+          // Disable retries, as useQueries
+          // will block successfull quotes from being returned
+          // if failed quotes are being returned
+          // until retry starts returning false.
+          // This causes slow UX even though there's a
+          // quote that the user can use.
+          retry: false,
+
+          // prevent batching so that fast routers can
+          // return requests faster than the slowest router
+          trpc: {
+            context: {
+              skipBatch: true,
+            },
+          },
+        }
+      )
+    )
+  );
+
+  // reduce the results' data to that with the highest out amount
+  const bestData = useMemo(() => {
+    return (
+      routerResults
+        // only those that have fetched
+        .filter((routerResults) => Boolean(routerResults.isFetched))
+        // only those that have returned a result without error
+        .map(({ data }) => data)
+        // only the best quote data
+        .reduce((best, cur) => {
+          if (!best) return cur;
+          if (cur && best.amount.toDec().lt(cur.amount.toDec())) return cur;
+          return best;
+        }, undefined)
+    );
+  }, [routerResults]);
+
+  const numSucceeded = routerResults.filter(
+    ({ isSuccess }) => isSuccess
+  ).length;
+  const isOneSuccessful = Boolean(numSucceeded);
+  const numError = routerResults.filter(({ isError }) => isError).length;
+  const isOneErrored = Boolean(numError);
+
+  // if none have returned a resulting quote, find some error
+  const someError = useMemo(
+    () =>
+      !isOneSuccessful && isOneErrored
+        ? routerResults.find((routerResults) => Boolean(routerResults.error))
+            ?.error
+        : undefined,
+    [isOneSuccessful, isOneErrored, routerResults]
+  );
+
+  return {
+    data: bestData,
+    isLoading: !isOneSuccessful,
+    error: someError,
+    numSucceeded,
+    numError,
+    numAvailableRouters: availableRouterKeys.length,
+  };
+}
